DB-5050: Add addWithDiff custom action
- Prepare removing test generators and templates from /src
- Moved test generators to __tests__, prepare to add tests for custom
  action
- Started adding next-wordpress-starter templates
- Add --ignore-path to lint-staged config
- Add ALL files in /src/templates to .prettierignore to prevent
  unwanting linting of templates

The custom action will need to be used any time we want the diff shown,
so likely "always". We can split out some of the logic if needed.
The type is overriding node-plop's type because when we abort, the
value of the promise is undefined, but plop wants a string.
Our custom type doesn't seem to be breaking the build but we should
consider fixing the upstream types at some point.