DB-6721:
- Fix some type errors in tests for the cli. There is
some more refactoring to be done here to make things a bit neater
or to make things more generic to help with the tests.
Switching to snapshot tests might help with some things also.
- Add typedoc tsconfig for the cli
- Update vitest to the latest version for all packages except the CLI.
Something with the new version of vitest broke a lot of the CLI tests,
will need more investigation. Left the version untouched for the CLI.
- Added an index.ts with exports for the health-check for typedoc