KIT-15:
Add a GitHub Action workflow that runs regression
tests on CLI generated starters on each push to
canary and nightly. Screenshots taken during tests
will be commited back to the repo via PR that is also
opened by this workflow.

A composite action that sets up playwright and
the monorepo was added as well. This is utilized in the
new workflow.